Fixed issue with Featured slider block
The slider container with not equal with the combined width of individual sliders wrapping 1 or two slides at the end which results to not being shown or breaking the layout.  This was supposed to be done automatically by bx slider but it doesn't work for some reason so we just coded something to calculate the corret width of tall sliders plus their margins.